rename bone to boneio, add support for oled
