+++ Code/BoneIO/MANIFEST.in
@@ -1,0 +1,4 @@
+include README.md
+include requirements.txt
+include LICENSE
+recursive-include boneio *.yaml
