--- Code/BoneIO/bone/version.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-__version__ = "0.1.0dev2"
