+++ Code/BoneIO/boneio/__init__.py
@@ -1,0 +1,11 @@
+"""Python library to control BoneIO."""
+
+from .version import __version__ as version
+
+name = "boneio"
+
+
+__all__ = [
+    "boneio",
+    "version",
+]
