+++ Code/BoneIO/boneio/version.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+__version__ = "0.1.0.dev7"
