+++ Code/BoneIO/create_dist.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+#source venv/bin/activate
+rm -rf dist/*
+python3 setup.py sdist bdist_wheel
+python3 -m twine upload dist/*
+
