+++ Code/BoneIO/example_config/output.yaml
@@ -1,0 +1,36 @@
+- id: light1
+  kind: mcp
+  mcp_id: mcp1
+  pin: 0
+  ha_type: light
+- id: light2
+  kind: mcp
+  mcp_id: mcp1
+  pin: 1
+  ha_type: light
+- id: light3
+  kind: mcp
+  mcp_id: mcp1
+  pin: 2
+  ha_type: light
+- id: light4
+  kind: mcp
+  mcp_id: mcp1
+  pin: 3
+  ha_type: light
+- id: light5
+  kind: mcp
+  mcp_id: mcp1
+  pin: 4
+  ha_type: light
+- id: light6
+  kind: mcp
+  mcp_id: mcp1
+  pin: 5
+  ha_type: light
+
+- id: switch1
+  kind: mcp
+  mcp_id: mcp2
+  pin: 0
+  ha_type: switch
